requirements.txt
Flask==3.0.3
yt-dlp==2025.01.08
flask-socketio==5.3.6
eventlet==0.36.1
python-dotenv==1.0.1
